[#15990] Fix CFPropertyList issues with OS X Leopard
Previously, Facter on OS X 10.5 (Leopard) was raising an error because
Facter::Util::CFPropertyList::List.load_str was returning a nil value.
This ultimately was because CFPropertyList was referring to the
LibXML::XML::Parser::Options constant, but that constant didn't exist
in the version of libxml-ruby that is shipped with Leopard.  This error
was fixed by rescuing the NameError due to the missing constant and
disqualifying libxml as a valid parser backend.  The second issue that
arose was a superclass issue around the XML class. This commit fixes
that issue by renaming the superclass from XML to LibXMLParser,
NokogiriXMLParser, and ReXMLParser in the backend files.